+++ src/main/java/org/cotato/csquiz/domain/recruitment/service/RecruitmentInformationService.java
@@ -5,10 +5,10 @@ import lombok.RequiredArgsConstructor;
 import org.cotato.csquiz.api.recruitment.dto.RecruitmentInfoResponse;
 import org.cotato.csquiz.common.error.ErrorCode;
 import org.cotato.csquiz.common.error.exception.AppException;
-import org.cotato.csquiz.common.schedule.RecruitmentScheduler;
 import org.cotato.csquiz.domain.generation.embedded.Period;
 import org.cotato.csquiz.domain.recruitment.entity.RecruitmentInformation;
 import org.cotato.csquiz.domain.recruitment.service.component.RecruitmentInformationReader;
+import org.cotato.csquiz.domain.schedule.RecruitmentScheduler;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.support.TransactionSynchronization;
