+++ src/test/java/org/cotato/csquiz/common/schedule/RecruitmentSchedulerTest.java
@@ -15,6 +15,7 @@ import java.util.concurrent.ScheduledFuture;
 import org.cotato.csquiz.common.util.TimeUtil;
 import org.cotato.csquiz.domain.recruitment.entity.RecruitmentInformation;
 import org.cotato.csquiz.domain.recruitment.service.component.RecruitmentInformationReader;
+import org.cotato.csquiz.domain.schedule.RecruitmentScheduler;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
