+++ src/test/java/org/cotato/csquiz/domain/recruitment/service/RecruitmentInformationServiceTest.java
@@ -15,10 +15,10 @@ import java.time.LocalDate;
 import java.util.List;
 import org.cotato.csquiz.api.recruitment.dto.RecruitmentInfoResponse;
 import org.cotato.csquiz.common.error.exception.AppException;
-import org.cotato.csquiz.common.schedule.RecruitmentScheduler;
 import org.cotato.csquiz.domain.generation.embedded.Period;
 import org.cotato.csquiz.domain.recruitment.entity.RecruitmentInformation;
 import org.cotato.csquiz.domain.recruitment.service.component.RecruitmentInformationReader;
+import org.cotato.csquiz.domain.schedule.RecruitmentScheduler;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
